tidy: Fix "multiplication implicit widening" in libmythfreesurround.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

1) Cast multiplication arguments to a larger type.

2) Convert variables to the appropriate type.

+++ mythtv/libs/libmythfreesurround/el_processor.cpp
@@ -421,8 +421,8 @@ private:
         }
     }
 
-    unsigned int m_n;                    // the block size
-    unsigned int m_halfN;                // half block size precalculated
+    size_t m_n;                          // the block size
+    size_t m_halfN;                      // half block size precalculated
     FFTContext *m_fftContextForward {nullptr};
     FFTContext *m_fftContextReverse {nullptr};
     // FFTs are computed in-place in these buffers on copies of the input
